requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.0
charset-normalizer==2.1.1
cycler==0.11.0
db-dtypes==1.0.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
google-api-core==2.10.1
google-auth==2.11.1
google-auth-oauthlib==0.5.3
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
imbalanced-learn==0.9.1
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.8
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.22.1
protobuf==4.21.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.2.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.8.0
